requirements.txt
numpy
Pillow
scipy
progressbar